made the extractable ocaml code compiled. 1) defined sz, align, INT as different modules with enough operations, such that the module generativity can ensure that we wont misuse them with nat or Z.
2) fixed two issues of extraction (fixextract.py)
  a) if you have two modules in Coq, 
   Module Type A.
     Variable a: Set.
   End A.
   Module B <: A.
     Definition a := nat.
   End B.
   Module C <: A.
     Definition a := Z.
   End C.
  With realizing B.a and C.a to be different types
  The extracted Ocaml *.mli is
   Module A = Sig 
     type a 
   End A
   Module B : A.
   Module C : A.
  although *.ml assigns the different concrete types to B and C.
  b) used Obj.magic to force Ocaml to type check monad join that is too
polymorphic.

3) extracted Ocaml into separate modules, rather than a large file.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@2685 eef5ea85-cc2e-0410-bd84-0fae1f012da8

+++ src/ssa/extraction/Camlcoq.ml
@@ -43,7 +43,7 @@ let camlint64_of_z = function
   | Zpos p -> camlint64_of_positive p
   | Zneg p -> Int64.neg (camlint64_of_positive p)
 
-let camlint64_of_coqint : Integers.Int64.int -> int64 = camlint64_of_z
+let camlint64_of_coqint : Integers.Int.int -> int64 = camlint64_of_z
 
 let rec camlint_of_nat = function
   | O -> 0
@@ -81,7 +81,7 @@ let z_of_camlint64 n =
   if n > 0L then Zpos (positive_of_camlint64 n)
   else Zneg (positive_of_camlint64 (Int64.neg n))
 
-let coqint_of_camlint64 (n: int64) : Integers.Int64.int = 
+let coqint_of_camlint64 (n: int64) : Integers.Int.int = 
   (* Interpret n as unsigned so that resulting Z is in range *)
   if n = 0L then Z0 else Zpos (positive_of_camlint64 n)
